+++ tempest_fastapi_sdk/__init__.py
@@ -355,7 +355,7 @@ from tempest_fastapi_sdk.websockets import (
     make_websocket_router,
 )
 
-__version__: str = "0.84.0"
+__version__: str = "0.85.0"
 
 __all__: list[str] = [
     "BASE_COLUMN_ORDER",